requirements.txt
PyYAML==6.0
docker-registry-client==0.5.2
requests==2.29.0
jsonpath-rw==1.4.0
pytest==7.3.1